// frontend/src/app/model/hotel.interface.ts
export interface Hotel {
    id?: number
    name?: string;
    username?: string;
    email?: string;
    openhour?: string;
    closehour?: string;
    reserveddate?: string;
    reservedcarnumbers?: string[];
    password?: string;
    role?: string;
    profileImage?: string;
    // passwordConfirm?: string;
  }
  